requirements.txt
# ICDEV Framework Dependencies
# Install: pip install -r requirements.txt

# === Core (required) ===
pyyaml>=6.0
jinja2>=3.1
flask>=3.0

# === BDD/TDD ===
pytest>=8.0
pytest-cov>=5.0
behave>=1.2

# === Code Quality ===
ruff>=0.12               # Ultra-fast Python linter (replaces flake8+isort+black)
pydantic>=2.0            # Test data validation (TestResult, E2ETestResult models)

# === Compliance ===
# cyclonedx-bom>=4.0     # SBOM generation (install when needed)

# === Search & Embeddings ===
openai>=1.0              # For embeddings (semantic search)
python-dotenv>=1.0       # .env file loading
numpy>=1.26              # Embedding math
rank_bm25>=0.2           # BM25 keyword search

# === Security Scanning ===
# bandit>=1.7             # Python SAST (install per project)
# pip-audit>=2.6          # Dependency auditing (install per project)
# detect-secrets>=1.4     # Secret detection (install per project)

# === Infrastructure ===
# boto3>=1.34             # AWS SDK for Bedrock/GovCloud (install when needed)

# === Monitoring ===
requests>=2.31           # HTTP client for ELK/Splunk/Prometheus APIs

# === SaaS Multi-Tenancy (Phase 21) ===
psycopg2-binary>=2.9     # PostgreSQL adapter for per-tenant databases
PyJWT>=2.8               # JWT validation for OAuth 2.0/OIDC auth
cryptography>=42.0       # RSA license keys, artifact signing, TLS
gunicorn>=22.0           # Production WSGI server for API gateway
redis>=5.0               # Rate limiting, session cache, pub/sub
